Add allocator benchmarks
* Added benchmarks for the Rho allocator in bench/allocbench.
* Added a benchmark script to run the allocator benchmarks:
  bench/allocbench.py.

The allocator benchmarks are intended to exercise the Rho allocator by
allocating very many objects, forcing stack scanning, and exercising the
garbage collector by allocating many objects without manually freeing
them.
