+++ bench/allocbench/huge-reuse.R
@@ -1,0 +1,2 @@
+dyn.load('allocator_test.so')
+.C('alloc_intvec', as.integer(10000), as.integer(20), as.integer(100000))
