+++ bench/allocbench/huge.R
@@ -1,0 +1,2 @@
+dyn.load('allocator_test.so')
+.C('alloc_intvec', as.integer(100), as.integer(100000))
