+++ bench/allocbench/large-lookup.R
@@ -1,0 +1,2 @@
+dyn.load('allocator_test.so')
+.C('alloc_lookup_intvec', as.integer(1000), as.integer(50000), as.integer(1000))
