+++ bench/allocbench/medium.R
@@ -1,0 +1,2 @@
+dyn.load('allocator_test.so')
+.C('alloc_intvec', as.integer(50000), as.integer(50))
