+++ bench/allocbench/small-lookup.R
@@ -1,0 +1,3 @@
+dyn.load('allocator_test.so')
+# Do very many pointer lookups.
+.C('alloc_lookup_scalar', as.integer(100000), as.integer(1000))
