+++ bench/allocbench/small.R
@@ -1,0 +1,2 @@
+dyn.load('allocator_test.so')
+.C('alloc_scalar', as.integer(100000))
